Update to main testing file
I have added slightly better documentation to the file as well as including ephemeral commands which prevent other guild users from seeing the command use. This is only mandatory for the "bug_report" command but the "time_difference" command has the option for a user to use the command privately.

Signed-off-by: Randy <rlbjr@mail.com>