Change to token implimentation
I have changed how the token is implemented into the code. Inside of hard coding the token into the code I placed it in a file and had python read that file. For some reason, probably where I have all my files stored, trying to get the current working directory gives an incorrect directory so as a temporary fix I just hard-coded the directory into the code. This can be seen in line 36 of BananaVariables. I will likely have to change my working directory in order to avoid this problem. If you attempt to run the code please change this directory as well as in lines 200, and 220.

Signed-off-by: Randy <rlbjr@mail.com>